Fix failing post submit test on API 24: Nexus Player
Several unit test fail because this device does not have
camera. So it will leave the setUp() method by assumeTrue()
which means CameraX.init() won't be called. But in tearDown()
method, it will CameraX.unbindAll() without CameraX.init()
and cause crash.

The solution is to check the initialization of CameraX before
calling CameraX.unbindAll() on tearDown() method.

Bug: 142378666
Test: ./gradlew camera:camera-*:test; ./gradlew camera:camera-*:connectedAndroidTest
Change-Id: Ibfb85f6340efbc67762e9f3e9b677e06ac302dac