Make Draggable children consume value lazily
This CL makes StateDraggable (previously was an effect) to provide
children with ValueHolder instead of real value, so children are now
able to read this value when they want and avoid being recompose on
every drag.

This way Drawer is not relayout-only primitive, which solves a lot of
lagging problems when the app content is a little bit more complicated
than just blank screen.

Bug: 142531095
Test: all Drawer and Switch tests passed
Change-Id: I67dcd427b63462526988aa22ad5a65cf6cc25c9a