Add Fragments in FragmentContainerView constructor
When adding inflated Fragments using FragmentContainerView, Fragment
currently do not get added until the view is attached to the window.
This is after the onActivityCreated callback, but it is desired that
developers can at least find the Fragment in onCreate.

This change adds a method on FragmentTransaction that takes a ViewGroup
so that we can execute a FragmentTransaction with a commitNow() in the
constructor of FragmentContainerView. This means the Fragment will be
available in the FragmentManager immediately after setContentView() is
called in the Activity. A check for finding the ViewGroup given in the
FragmentTransaction was added to FragmentStateManager, so if a ViewGroup
is present, it does not check for the containerId.

Test: added FragmentContainerInflatedFragmentTest, tested in nav app
BUG: 142520327
Change-Id: Id14787dff45f19348a8d423a5b20d82ab606ec6a