Add AnnotatedString utility constructors
Even our tests used lots of examples where a single TextStyle or a single
ParagraphStyle is added to an AnnotatedString. 

Added utility functions to help developers easily construct an AnnotatedString
using a single TextStyle and/or ParagraphStyle which will be applied to the
whole text.

Moreover, switched tests and samples to use new functions and 
AnnotatedString.Builder in places where it made sense.

Test: ./gradlew ui:ui-text:test
Test: ./gradlew ui:ui-text:connectedAndroidTest
Test: ./gradlew ui:ui-framework:test
Test: ./gradlew ui:ui-framework:connectedAndroidTest
Test: Manual check on Demo app

Bug: 139312718
Change-Id: I1c72b908fc6564efe04623d58846bd1cc0f6f515