Ensure fingerprint dialog always dismissed on error
Currently, if FingerprintHelperFragment receives an error other than
ERROR_CANCELED or ERROR_LOCKOUT/ERROR_LOCKOUT_PERMANENT, it fails to
trigger the error callback to dismiss the dialog before the helper
fragment is cleaned up. This commit fixes this issue by moving the call
to cleanup() into the Runnable that's posted in this case. It also adds
a test case to FingerprintHelperFragmentTest to exercise this behavior.

Test: ./gradlew biometric:connectedAndroidTest
Fixes: 143683687
Change-Id: Ibe1ef02dd64e3fe2dac11496560241a1a3e7b735