Add Focus ComponentNodes
Add component nodes to implement focus as part of the component tree.
When a component needs focus, it makes a request to it's parent, and the parent either grants or denies the request.
Since we don't know about focusable children during composition, we compose the UI with an initial non-focused state, and then grant focus to the root node.

Bug: 143979358
Test: Added unit tests and integration tests.
Change-Id: I54abf138c73f92e06ff89e253303c617868571b7