Adds contentColor metadata to foundation
contentColor represents the color of typography and
iconography in a given subtree. This is set by Surface
to correspond to the `on` color if Surface's color is a color
in the theme.

If not, Surface will use the existing value of contentColor(),
which defaults to black if not set by any other component.

Bug: b/145355526
Test: SurfaceColorsTest

Change-Id: Icd0743ece7c29e24e8af341b47659425e30eb838