More on initial state
Transition composable needs to accomdate two different use cases:
1) Initial state is the state that the transition is currently in until the
state changes.
2) Initial state is not the current state. Its only purpose is to provide a
starting point for an animation that animates from an old state to
the current state. An exmaple of this could be a composable that animates
in when it gets added to the tree, or a progress indicator.

This CL adds an optional initState param to transition that defaults to the
first toState ever passed to the transition. This is the majority of the
useage (also the use case #1 list above). The initState param can also
be explicitly specified by devs to achieve the use case #2.

BUG: 144683881
Test: Infinite progress indicator working properly
Change-Id: I2d19a64ed6ae08b8ce4b53c5c6e86c547c40640d