Fix the models reads inside the composition and writes in WithConstrains measure block
During migrating some components to WithConstraints we faced an interesting incorrect behaviour.
There are cases when we modify some model inside onMeasure and then use this updated value inside composition. It works fine without WithConstraints as measuring and composition never were happening in the same frame. Now it is possible as WithConstraints is subcomposing inside the measure block and then performs the measuring, so this operations happens in the same frame. And the recomposition wasn't happening because the write to the model happens in the same frame where we created the model and such cases triggers a bit different and CommitFrameObserver will not be called with such model. To make this possible we should always switch frames and do measuring not in the same frame we do subcomposition.

Test: a test for this use case in WithConstraintsTest
Change-Id: I204fa600b19b7801b4ef20decc3ba10d79b02c10