Define swipe by fx(t) and fy(t) instead of p0 and p1
This opens up the possibility to dispatch arbitrary gesture paths to a
composable, by encoding it in fx and fy. To make sure it samples an
input event at specific coordinates and times, one can define key times
where an input event will be forced.

To be able to test this, the AndroidInputDispatcher now allows changing
the eventPeriod. To speed up the tests, the dispatcher is configured to
not wait until the event time before sending the event.

Bug: 143480427
Test: AndroidInputDispatcherSendClickTest,
      AndroidInputDispatcherSendSwipeLineTest,
      AndroidInputDispatcherSendSwipeFxFyTest

Change-Id: I8e7133ba0a3bf97e96d2e9fdf8c6cf191b483974