WindowManager Support Library - Samples
Initial commit for samples for Window support library.
This includes three sample activities:

1. Shows current device status, all display features reported for
the app window and the log of the changes in state.

2. Showcases SplitLayout - a view group designed to separate UI to
sides of a display feature that divides the window into two areas.
If there is no split, or if the split creates areas are not suitable
for child views (e.g. too small to fit minWidth/minHeight), then it
behaves like FrameLayout, stacking the children on top of each other.

3. Showcases tracking of the state of a foldable device and showing
some content on the outside screen using Presentation API.

Test: Manual, using sample app.
Change-Id: I5b71401a4520d14a07fe6f1d3333cb7981d52486