Remove PagingState
Test: ./gradlew paging:p-c:test paging:p-run:cC

PagingState was an exploration of a fully-state-resolved flow, in
which elements are immutable, and do not require any additional or
historical information to present in the UI.

The big upside of this approach is that if one downstream consumer
stops consuming for any reason (e.g. stopped activity in back stack)
but the producer keeps pushing, we don't need to buffer (unlike with
events).

The big downside though is that we cannot efficiently transform these
objects. Any transform on a PagingState would have to be re-performed
any time a page was loaded, or an loading state changed. This is
extremely difficult to do efficiently, since it's O(loaded dataset)
instead of as with PageEvents, where it's O(changed dataset).

Eventually, we realized we could duplicate its benefits with a caching
operator on a Flow<PagedData>, which could avoid the infinite
buffering problem.

Change-Id: I583225c779b756d7ed1699f3c90427ee153b6114