Integrate native fling spline into Compose UI
Extract the native Android fling spline from android.widget.Scroller
and move it into Compose UI foundation in such a way that it can be
driven via an external time source.

Future work will delegate to a similarly factored version of this
in AOSP frameworks/base for a future platform release for forward
compatibility with any future platform behavior changes.

Thanks to the nature of the source material, testing this port for
correct behavior will necessarily involve an instrumentation test and
comparing values over "real" time with the output of
android.widget.Scroller.

This commit also starts spreading a new-ish pattern of @Composable
factory functions that read default or configuration-specific values
from the composition. An interesting outcome of this is that the result
of such a factory may not be the content of a `remember {}` block - the
`remember` body is not itself `@Composable`. In order to preserve
expected object instance behavior as well as a reasonably usable caller
experience this means that such factories must internally `remember`
as needed.

This might be for the best. When this pattern is used for complex
hoisted state objects that require composition-based construction, it
removes the need for an explicit `remember {}` on behalf of the caller
and the right thing happens with simpler calling code. On the other
hand, this may end up being confusing for developers who may not easily
recognize when something needs to be `remember`ed.

Current builds of the Android Studio plugin highlight composable calls
in bright green and once an outstanding plugin bug is fixed, attempting
to memo the result of one will be a compiler error. This might be
enough. If not, we should explore other tooling (lint?) or conventions
(`rememberFoo` for such factories?) to help aid in understanding
and discoverability.

Relnote: Scrollers now exhibit native Android fling motion behavior.

Bug: 147493715
Test: existing; manual

Change-Id: I922af68261f3f1e81538a98a7575603e531fc035