Don't remove observers that are not subscribed
When an AnimationClockObserver is unsubscribed from an AnimationClock
before it is subscribed, it lingers in the toBeRemoved list. If it is
then subscribed before the next dispatchTime, it will be immediately
removed in the first dispatchTime invocation.

Fix this by checking if the observer should be added to the toBeRemoved
list, which is only if it is in the observer list.

Bug: 147736746
Test: ./gradlew ui:ui-animation-core:testDebugUnitTest \
        --tests AnimatedValueTest

Change-Id: I03d9ab54e1a5356ae4244849c362131fc739600d