Add CollectionItemInfos to ViewPager2
As of r.android.com/1093531, RV won't populate the CollectionItemInfos for its
children. VP2 doesn't have the same issues as RV that motivated the
change (all pages behave as part the collection, all are important to
a11y), so re-add the infos here.

Also added contentDescriptions to ImageViews in the demo app so TB focuses
and announces them.

Test: Added tests. Checke counts on demo app.
Change-Id: I39c2c6c89ffd848ef06b243db04ca5178122438d