Provide PreviewSurfaceProvider asynchronously inside Preview.
Currently, when binding a Preview use case to a camera, it is only attached to the camera id if its PreviewSurfaceProvider is set. This CL changes this behaviour so that the preview use case can be bound even before the PreviewSurfaceProvider is set, this will allow to call bindToLifecycle() before preview.setPreviewSurfaceProvider(), which currently results in no preview being displayed.

This will solve an issue in aosp/1187216, where the cameraInfo is needed in order to set up PreviewView.

Bug: 143703289
Test: manual testing, PreviewTest.java
Change-Id: I519fbca343f38a4ce6c9e8aa45fe6107dd7c784e