Test animation in an Activity
aosp/1089069 changes animation scale to 0 in settings for androidx test.
When activity is created, it reads the settings and write 0.0f to ValueAnimator.sDurationScale.
(see android.view.WindowManagerGlobal.getWindowManagerService)

Because of that, for the test that do want the animation duration, it must change the value to 1.0
*after* activity is created.

Existing AnimationDurationScaleRule doesn't work for the case when
ActivityTestRule is created with "launchActivity=false". The activity is
created after the rules are applied.

This change created a subclass of ActivityTestRule, reads
"@AnimationTest" flag in apply() then change the scale right after
activity is created (both launchActivity=false/true)

The change makes GridWidgetTest#test27766012 working again.
Also fixed few couple other tests relying on animation.

Bug: 141172007
Test: /gradlew leanback:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.leanback.widget.GridWidgetTest#test27766012
Change-Id: Ia98788c2e6b28f0d6ec109472dd4d4d4a97a81c3