Free AnimatedValue from ValueHolder. @Composable animate.
Remove the ValueHolder class and AnimatedValue's dependency on it. Instead, AnimatedValue
now keeps an abstract value field with a protected write. AnimatedWhateverModel's can then
be created to track the change of that value field for the purpose of properly recomposing
whoever reads that value. This effectively achieved the goal of ValueHolder while maintaining
an elegant solution. It also allows for future support of View property animations
with AnimatedValue.

Another change in this CL is the introduction of animate(..) composables. They are light-
weight AnimatedValues - We wrap the AnimatedValue in a composable that takes a target
value and returns a current value. While taking away the ability to stop/cancel the animation,
fling, etc, we get a largely simplified API for animating in-between predefined snapshot
values. Depending on the user feedback, this line of composables may be expanded to support
all compose built-in data types.

RelNote: "Removed ValueHolder class. Restructured AnimatedValue, AnimatedFloat classes to
make the animation value field abstract so that subclasses can watch the value update.

Added model classes for AnimatedValue, AnimatedFloat, etc.

Added a new set of light-weight @Compopsable APIfor animating between values."

Test: new api demo

Change-Id: I79530e117cfa893a52542f85a55528eaa0f11b93