Do not calculate selection for the detached Text layout coordinates.
In the text selection logic we store the layout coordinates object associated with the Text component inside TextSelectionDelegate. In most cases everytime this layout is detached we should also unsubscribe the TextSelectionDelegate in this case, but due to an asynchronous nature of all this callbacks there could be the cases where the selection is being calculated for recently detached text layout. To not crash in this cases we will work with a detached text layout as containing an empty selection.

Bug: 148048619
Test: manually, wasn't able to cover with a stable test
Change-Id: I440cdace1bc2f13e7780d24c34786f5088139052