API Change: Support URI and OutputStream in ImageCapture
Add APIs in ImageCapture to support URI and OutputStream. For URI, the caller is responsible for specifying the collection of the MediaStore and the ContentValues. The cameraX is responsible for manipulate the IS_PENDING flag on top of it. For OutputStream, the caller is responsible for opening and closing the stream.

- New APIs in ImageCapature.
- Refactored the overloaded takePicture() calls to accept a reusable OutputFileOptions.
- In OnImageSavedCallback, return the saved URI if available.
- Update ImageSaver to accept OutputFileOptions. For Uri and OutputStream, a temp file is first created before writing to the final destination. This is neccessary because ExifInterface only supports File.
- Update test app to use Uri as the canonical example for image capture.

Relnote: "Updated ImageCapture to allow saving images to Uri and OutputStream. Combined over loaded takePicture methods into one. Updated test app to use Uri as the canonical example."

Bug: 148305437
Test: manual test and ./gradlew bOS
Change-Id: Ia3becf13029c367e7525d1b110f89bd4a06833a9