Migrates Material ambients to staticAmbient
All these ambients (with the exception of ColorAmbient)
change infrequently and are read a large amount of the time, so changing
these to static ambients improves read performance.

For ColorAmbient specifically, given that ColorPalette contains a lot of
properties, even in the case of a normal ambient writes to one property
will cause invalidations across huge parts of the application - so the
recommended pattern here is instead memoize an object that contains
State backed properties, so only reads to one specific property get
observed. In this case, since we memoize the main ColorPalette and don't
anticipate this object changing, rather only mutating properties inside,
making ColorAmbient also static is both safe and gives us notable wins
in read performance in components.

Also fixes ColorPaletteBenchmark, which was broken after the Ambient
refactor.

Test: ColorPaletteBenchmark
Change-Id: I589b44b32f88e3e752bc65b333270c0a68d363e6