Try to fix the test failure on 21 and 23
I didn't manage to reproduce this failure on emulator and a real device both running 21 so looks like it is not api level related. Looks like this devices are too slow and the assertion for a countdown latch to be fired during 200 ms is too strict for them. I fix this to wait 1 second in cases when we do expect the latch to be fired.

Bug: 148004999
Test: tests are passing on api 21 and 29
Change-Id: Id2cfe70790f0dfdf98d10c4fd0f5e3acef686278