Disable clipChildren for GhostViewPost
There is a bug in the ported version of GhostView used by ChangeTransform(all platform levels apart from P). The view added into GhostView can be not drawn at all even when it should be drawn. The platform implementation which we use on P works fine. Why it happens: sometimes we apply an animation matrix for a view added into GhostView. This means the view will not be drawn inside their nominal bounds - [mLeft, mTop, mRight, mBottom]. If GhostViewPort has clipChildren() == true then because of Canvas.quickReject(mLeft, mTop, mRight, mBottom) returning true the drawing of such child would be skipped, even if in fact this child is visible due to the animation matrix transformation. Disabling clipChildren is fixing the issue. There is no clipping applied for a ghost view in the platform implementation as well.

Test: manually on the project provided by an external developer and via a new test
Bug: 148798452
Change-Id: I9b9ceddf8e8bcc6e475b00ea10275cac1274ce78