Fixes SampledAnnotationEnforcer looking through all children
This was causing resolving the wrong sample directory - there are only
two places for the sample directory to exist so we should just look
there instead of recursively searching the tree.

Unfortunately it turns out that most of the tests were only accidentally
passing, as submodules arent' actually supported when testing lint and
the detector was accidentally finding the package directory, not the
actual samples directory (that happened to also be named samples).
Removed tests that we cannot verify so at least we can still ensure most
of the functionality works.

Test: ./gradlew ui:ui-material:lintDebug
Change-Id: I280ce2e3536554dcc22da657443aec768ab27c80