App bar simplification
Providing built-in support for the number of actions available in app bars seemed like a nice idea at the time, but it ended up being restrictive and resulting in an overly complicated API surface for both the simple case, and anything morecustom.

This CL removes the concept of 'actionData' and 'action', greatly reducing complexity of the majority of use cases.

Handling different orientations of start / end icons in a Scaffold when the FAB position changes is also fairly simple, as you can just use the fab position directly and change the order of items emitted into the BottomAppBar.

Overall I think this makes app bar APIs easier to understand and also more powerful.

Bug: b/148049128
Test: AppBarTest
Change-Id: I5bfee975b85a674ace23a6f256492f34aa9229ea