Handle re-entrant calls to moveToExpectedState()
As moveToExpectedState() runs user code that can
potentially change the external state, it is possible
for re-entrant calls to moveToExpectedState() to
happen. In this case, we can ignore the re-entrant
call, knowing that after the current state transition
completes FragmentStateManager will call
computeExpectedState() after each transition and pick
up the updated state.

Test: this isn't technically possible yet, but it will be
BUG: 139536619
Change-Id: I500e6f0e78d20355ab21a1f21fc66067f72e46a9