Fixes deferred and imminent recompositions
These were accidentally broken during a refactor, as a result of this change ColorPaletteBenchmark also works again.

Also modified ColorPaletteTest to properly ensure the number of recompositions at the end of the test.

The tests guaranteeing this logic are un-ignored in AmbientTests, and we will track them to see if they continue to flake in the future.

Bug: b/149912466
Bug: b/148810324
Test: AmbientTests
Change-Id: If2c861a951d1ee58c44d042acac1a2f88796d4f1