Disposable animation clock
AnimationClockObservable has a gotcha - it's possible to leak
subscriptions to it as it isn't intrinsically scoped. This change adds
a composable .asDisposableClock() extension to AnimationClockObservable
that throws away subscriptions after the call leaves the composition.

A suspending API similar to the awaitFrame() function available in
kotlinx.coroutines.android provided in some sort of ambient object may
be the better path forward, but for the time being this addresses the
gotcha in the existing API surface.

Test: existing animation tests, DisposableAnimationClockTest.kt
Change-Id: I7a50bf7c691c262435eca8c8dff80c75cbd11c45