Changed bindToLifecycle to modify UseCase only if success bind occurs
This changes CameraDeviceSurfaceManager to operate on the SurfaceConfig
and UseCaseConfig to do resolution calculations instead of operating on
the UseCase instance.

In addition added the onUnbind() method to UseCase which gets called
when a UseCase is unbound. This is the counterpart to onBind() which
gets called when a UseCase and Camera are bound together.

Removed attachCameraControl() and detachCameraControl() from UseCase
since the CameraControl is available once the Camera is bound to the
UseCase.

Removed UseCaseTest.clearListeners() since the clear() method should
only do UseCase specific clearing, while the onUnbind() will clear out
the listeners.

Relnote: "Fixed bindToLifecycle so that it only modifies the UseCase if
it successfully binds. Previously when calling bindToLifecycle in order
to do resolution calculations the UseCase is updated. Now it no longer
needs to update the UseCase to do the calculations"

Changes in SupportedSurfaceCombinationTest to use Truth.assertThat
instead of JUnit.assert* methods.

Bug: b/147698557
Test: ./gradlew camera:camera-{core,camera2}:test
      ./gradlew camera:camera-{core,camera2}:connectedCheck
      manual test of core-test app

Change-Id: I78d9ec96d30e2a2487ff43119f8b2067c8b92a03