Initial commit of Typeface
The Typeface represents a loaded FontFamily. There are three Android
implementation based on FontFamily type.
This will be a replacement of TypefaceAdapter which adaptively loads
font on UI thread. Instead, Typeeface will be loaded in background
thread as the same way of async resource loading like VectorDrawable.

This CL only introduces Typeface implementation and test case.

Bug: 143715852
Test: ./gradlew test
Test: ./gradlew :ui:ui-text:connectedAndroidTest
Change-Id: I37befd6d1297acd6485c1c778c3e5be601619b47