Move makeInactive() out of moveToState()
Instead of calling makeInactive() as part
of moveToState() / moveToExpectedState(),
run that logic *after* the state settles.

Updating the target fragments correctly can
then be moved solely to FragmentStateManager's
destroy(), ensure it runs at the right time.

Special care had to be made when restoring
retained fragments that aren't in mActive so
that they are manually removed from the mNonConfig
(as before that would have been take care of
by moveToExpectedState()).

Test: all existing tests pass
BUG: 139536619
Change-Id: Ie33e01622f5edf212822f9c75f60c532a352a2d1