Make onActivityCreated optional in DialogFragment
DialogFragment uses onActivityCreated() to help setup the dialog. Since
onActivityCreated() is being deprecated (aosp/1189651) we should find a
better home for the logic executed there.

Code concerning the view has been moved to the observe of the
viewLifecycleOwnerLiveData of the DialogFragment, and other
initialization can be done in onGetLayoutInflater when the Dialog is
initialized.

We will still support dialog setup in onActivityCreated for custom dialogs.

Test: all existing tests pass
BUG: 144309266
Change-Id: I832c596f2a6d6c2a4d77b97df15930d39bdedb9d