Updates EmphasisLevels implementation to use luminance of the content color
Previously we set the implementation specifically for `onPrimary` and `onSurface` backgrounds, but this solution does not scale to other colors, such as onSecondary.
Instead now EmphasisLevels are implemented in terms of the luminance of the incoming color, and whether the theme is light or dark. This allows setting the correct alpha and hence contrast / accessibility for any incoming color.

Bug: b/145354279
Test: EmphasisTest
Change-Id: I048890f266b53bcb76bf88fbfac1f4dc6d16b8c8