Introduce savedInstanceState() and rememberSavedInstanceState()
These functions are similar to state() and remember() but have a saved instance state support build in.
There is also a new Saver class which allows to define how to save the complex class by converting it into something which is supported by Bundle.
Plus I added a new test api which we will use to test how our components survive the state restoration.

Relnote: New functions: savedInstanceState() and rememberSavedInstanceState(). They are similar to state() and remember() but have a saved instance state support build in
Test: manually, new tests
Bug: 152025209
Change-Id: If1987a758d18f9fa4ccfeb75011155304ee99cef