Make most Activity Result APIs final
APIs not meant to be extended should be
made final.

For the pre-built contracts in
ActivitResultContracts, most of them have
had a comment added to explain that they can
be extended to allow for passing additional,
optional extras to the Intent created by
super.createIntent(). If additional functionality
(or changing the parsing logic) is needed,
it is still recommended to create a custom
contract.

Test: ./gradlew checkApi
BUG: 152439361
Relnote: "Many of the Activity Result not meant to
be extended are now final."

Change-Id: I6a2ced62b158c2bbda6ca5ce5d16078d6a491d42