Update FocusManager to be able to observe focus transition other than FocusNode
I'm pushing click-to-focus code to ui-foundation. To be able to observe
the focus transition in CoreTextField, split FocusNode and observer.

This also change the callback interface from onFocus/onBlur to
transition which is more general than simple onFocus/onBlur without
argument. With this callback, observer can know the next focus client
and determine the software keyboard needs to keep open or not.

Bug: 152566483
Test: TreeHugger
Change-Id: Iffcd42876362c991912f8e86ab096dd7b0c5cbc3