Simplify api of RippleEffect and RippleEffectFactory
We don't need requestRedraw param on RippleEffectFactory.create() as we can trigger redraw by using MutableState directly in RippleEffect as an implementation detail.
We will keep RippleEffect.dispose() to be flexible as other implementations could need it, but we stop using it in DefaultRippleEffect as the disposing will be done automatically as we use DisposableAnimationClock. Added a test for it.

Fixes: 137183289
Test: RippleTests are passing, covered with a new test
Change-Id: I82f21e767a3da81961e2d020fe538aef48f3c3b4