Add support for passing results between Fragments
Currently if developers would like to pass data between fragments that
are in the same FragmentManager, they use the target fragment APIs.
Using target fragments requires direct access to a fragment instance,
which can be dangerous since the state of the target fragment is
unknown.

This change offers APIs to set a result on a Fragment and handle how
that result is used at the proper time in the Fragment's lifecycle. The
same usage pattern will work for passing data between any two fragments,
including DialogFragments, and fragments used in Navigation.

This change also includes a -ktx extension function to ensure kotlin
users can pass in the FragmentResultListener as a hanging lambda.

Test: added FragmentResultTest and FragmentResultOwnerTest
Bug: 149787344
RelNote: "Added support for passing results between two Fragments. This
works for hierarchy fragments (parent child), DialogFragments, and
fragments in Navigation."

Change-Id: I0d9a1f1a66b872eb5ca08f132e4f7838abaad962