Ensure getContents and OpenDocuments return single elements
When using the getContents and OpenDocuments contracts if you selected
multiple items it worked corrected, but selecting only one item returned
null. This is because the intent stores multiple items as `clipData`, but
a single item as `data`.

This change ensures that we handle the `data` so selecting single items
now works.

Test: added verification to integration app
Bug: 152941153
RelNote: "When using the getContents and OpenDocuments
ActivityResultContracts and selecting a single item, it is now correctly
returned to your callback."

Change-Id: I92b445feaf7fc651cdf9e7787a84e839d723da00