Add test and temporary fix for padding placement issue in Rtl
There is an issue where the padding will not correctly position the wrapped layout when it is placed in between layout direction modifiers. This Cl adds a test for this use case. Also it temporary falls back to a previous logic where we manually handle the rtl placement in padding modifier by checking the layout direction and calling placeAbsolute with calculated position.

Bug: 153317665
Test: new test and old tests passed
Change-Id: Ib1ef45d7e67721d6c1941700bbecb9f8ecf4604c