Fix more cases of inline classes cross module.
The previous fix only took care of accesses to the main underlying
value of the inline class. However, other getters on the inline
class also need to be mangled. In order to make that work we
need to make sure that properties and getters are properly hooked
up also when used across module.

Bug: 152232807
Test: KtxCrossModuleTests.testInlineClassCrossModule
Change-Id: I01ae0a01e27ba174c83de8d83c8539f0a893dc4d