Add DND fail detection in camera2 compat module
A RuntimeException will be thrown by satisfying the following conditions
* An early Android P(28) release
* Camera hardware level is INFO_SUPPORTED_HARDWARE_LEVEL_LEGACY
* Device Do Not Disturb (DND) mode is enabled
* Invoke API CameraManager.getCameraCharacteristics() or CameraManager.openCamera()

The CL contains
(1) Create a CameraAccessExceptionCompat, it defines DND fail error code combining with original CameraAccessException error.
(2) Catches the DND RuntimeException and rethrow a CameraAccessExceptionCompat.
(3) It caches all CameraCharacteristics in API 28 and returns the cached one only when running into DND fail case. It is a workaround to lower the chance to encounter DND RuntimeException.

Bug: 149413835
Test: ./gradlew camera:camera-camera2:test
Change-Id: Ia04a9cfe61fca9e61dd02ea5f44f6bbb771e9a53